Use GHC header parser for module name extraction
Replace ad-hoc word-based extractModuleName in Scrod.Cabal with
GHC's Header.getImports, which correctly handles comments and
pragmas between "module" and the module name. Moves the function
to Scrod.Ghc.Parse following the existing discoverExtensions pattern.
